넓이 계산하는 함수 프로토콜 분리
