+++ CoordinateCalculator/CoordinateCalculator/InputView.swift
@@ -9,14 +9,14 @@
 import Foundation
 
 struct InputView {
-    
+
     func readInput(_ message: String) -> String {
         print(message)
         let input = readLine()
         guard let inputCoodinates = input else { return "" }
         return inputCoodinates
     }
-    
+
     //(10,10)-(20,20) "-" 기준으로 분리
     func makePoints(_ input: String) -> [MyPoint]? {
         var coordinates: [MyPoint] = []
@@ -26,17 +26,17 @@ struct InputView {
         }
         return coordinates
     }
-    
+
     //(10,10) "," 기준으로 분리
     private func separateComma(_ input: String) -> Array<String> {
         let coodinates = input.components(separatedBy: ",")
         return coodinates
     }
-    
+
     private func createPoint(_ input: [String]) -> MyPoint? {
         guard let first = Int(input[0].dropFirst()), let last = Int(input[1].dropLast()) else { return nil }
         return MyPoint.init(x: first, y: last)
     }
-    
+
 }
