+++ CoordinateCalculator/CoordinateCalculator/MyCalculate.swift
@@ -1,0 +1,15 @@
+//
+//  MyCalculate.swift
+//  CoordinateCalculator
+//
+//  Created by 김수현 on 2018. 1. 30..
+//  Copyright © 2018년 Codesquad Inc. All rights reserved.
+//
+
+import Foundation
+
+protocol MyCalculate {
+
+    func calculate() -> (String, Double)
+
+}
