+++ CoordinateCalculator/CoordinateCalculator/MyLine.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-struct MyLine: MyShape {
+struct MyLine: MyShape, MyCalculate {
     
     var pointA = MyPoint(x: 0, y: 0)
     var pointB = MyPoint(x: 0, y: 0)
