+++ CoordinateCalculator/CoordinateCalculator/MyShape.swift
@@ -9,10 +9,10 @@
 import Foundation
 
 protocol MyShape {
-    
+
     func drawPoint() -> [MyPoint]
-    
-    func calculate() -> (String, Double)
-    
+
 }
+
+
